# Improve UTCTime encoding