The encodings given by the CBOR RFC are either inefficient (a
ISO8601 string) or imprecise (offset from the epoch in seconds).
Here we implement the extended time RFC (currently an [Internet Draft](https://tools.ietf.org/html/draft-bormann-cbor-time-tag-00#section-3.3)) which provides a greatly improved encoding.

Currently the draft lacks an assigned tag, so for the time being I've stolen `0x2221feca` (chosen at random via roll of a very large dice). I've started the IANA process for getting a tag assigned while the RFC process takes its course. Unfortunately, eventually updating the library to reflect the assigned tag will be a breaking change.

## Criterion Results

Benchmark | Before | After
------------|------------|-------
serialise UTCTime | 969.9 μs | 65.88 μs
deserialise UTCTime | 1.316 ms | 112.0 μs