@bgamari 2017-02-23
> We can't change the existing instance anyway without breaking existing code, so this needs to be refactored anyway.

@thoughtpolice, I don't see how this would break existing code. We can read all of the encodings that we could previously. The only foreseeable issue is someone trying to read a file with a tag 40 encoding with an older version of the library. However, seeing as this isn't released yet I really don't see this as a problem. Afterall, they merely need to recompile their project against a newer `binary-serialise-cbor` to regain compatibility.