@bgamari 2017-02-26
@thoughtpolice ping?