@sopvop 2017-02-27
What about users who use CBOR to communicate with non haskell programms?