@bgamari 2017-02-27
Ahh, I neglected to mention that I am working to get the encoding proposed here (or something like it) granted an IANA-sanctioned tag. Moreover, the IETF has created a working group which will be taking CBOR to a full Internet standard; there is a fair chance that we will be able to get this encoding in that specification.