@thoughtpolice 2017-02-28
@bgamari Ah, no, you're correct. I misread the tag, so this is absolutely my fault because I'm an idiot.

The `docs` file should actually go in the Tutorial module or something I think, though; just having the documentation on the instance itself is rather out of line. But otherwise yeah, this seems OK.