@axman6 2017-03-03
I think my comments may have got lost/not seen because some of the code changed, but:

> Pretty sure tag 2 shouldn't be there in the error message (at least, it doesn't look like the code supports a tag of 2, unless we're supporting BigNum as a representation of time too).