@bgamari 2017-03-04
> I think my comments may have got lost/not seen because some of the code changed, but:

Thanks for the reminder, @axman6.