@bgamari 2017-03-14
For the record, we are currently in the final stages of submitting a final encoding for consideration by IETF. The final version is quite a bit different then what I have implemented here; I'll update this branch shortly.