@lfairy 2017-03-25
I found a draft of the proposal here: https://tools.ietf.org/html/draft-bormann-cbor-time-tag-00

Not sure how up to date it is, though.