@thoughtpolice 2017-03-25
Yes, it was published on the 13th of March (look below the author field) and the draft expires before September 23rd (a period during which revisions can be sent in and before another one has to be made). So I believe this has the differences Ben mentioned earlier (but he hasn't updated this PR)

(The IETF draft setup is a bit weird like that :)