@axman6 2017-03-26
It looks like that proposal has grown _a lot_ from the initial proposal here, is there a good reason for it to be so complex? Also reading the RFC, it would be good if there were examples of all the proposed structures. (making the comment here because I have no idea how actually RFC comments are supposed to work!)