@dcoutts 2017-03-26
> It looks like that proposal has grown a lot from the initial proposal here, is there a good reason for it to be so complex?

@axman6 You're of course welcome to contribute your feedback to the RFC process. I'd guess it's other feedback from other people about things they want.

But note that (unless I've missed something) we don't need to make use of all the flexibility available. It's worth looking specifically (and e.g. documenting here for the rest of us to review) what of the available options we would use, what that encoding would be exactly and how big it would be.