@bgamari 2017-04-12
Alex Mason <notifications@github.com> writes:

> It looks like that proposal has grown _a lot_ from the initial
> proposal here, is there a good reason for it to be so complex? Also
> reading the RFC, it would be good if there were examples of all the
> proposed structures. (making the comment here because I have no idea
> how actually RFC comments are supposed to work!)
>

This proposal isn't Haskell-centric; I'm merely one author. The proposal
is intended to address the general inadequacy of the date/time encodings
laid out in the original CBOR specification so yes, it is a bit more
complex than we strictly need.