@bgamari 2017-06-13
This should be ready for review.