@axman6 2017-06-14
Just confirming, this change doesn't implement the whole RFC, but only the bit necessary to efficiently encode UTCTimes?

I guess this should the optimised common case for encoding and decoding in Haskell, but maybe we should also provide a Times module which supports the whole RFC as both a proof of concept and to make life easier for those in the future need to interact with foreign code which does implement other parts/all of the RFC.