@bgamari 2017-06-14
> What do the criterion numbers look like now?

See PR summary.