@dcoutts 2017-06-14
No objection to providing greater support for the date/time RFC. PRs along those lines would be accepted.

Criterion numbers look good now.

Anything left to do or can we merge?