@bgamari 2017-06-15
We now have an official IANA sanctioned major tag, 1000!

I think this is ready to be merged.